# Querier: Reduce allocations on `QueryStreamResponse` unmarshal