… pkg/ingester/client/wrapped.go