Signed-off-by: Miguel Ángel Ortuño <ortuman@gmail.com>

<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

Use custom query stream response unmarshal code to allow reusing chunk slices when querying ingesters. 

To test the changes, new `BenchmarkQueryStreamResponseUnmarshal` has been added and executed with the following command: 

```sh
go test -bench=^BenchmarkQueryStreamResponseUnmarshal$ -run=X -cpu=8 -count=10 ./pkg/ingester/client
```

Below are the results for benchmark comparison between the previous and the new implementation: 

```
name                                                               old time/op    new time/op    delta
QueryStreamResponseUnmarshal/10_series,_10_chunks,_100_samples-8     51.6µs ± 1%    33.6µs ± 0%  -34.91%  (p=0.000 n=9+9)
QueryStreamResponseUnmarshal/10_series,_100_chunks,_100_samples-8    76.5µs ± 1%    41.6µs ± 0%  -45.61%  (p=0.000 n=10+9)
QueryStreamResponseUnmarshal/100_series,_1k_chunks,_1k_samples-8     4.43ms ± 0%    2.48ms ± 1%  -44.07%  (p=0.000 n=9+9)
QueryStreamResponseUnmarshal/100_series,_10k_chunks,_1k_samples-8    26.7ms ± 0%    11.5ms ± 1%  -56.68%  (p=0.000 n=9+10)
QueryStreamResponseUnmarshal/1k_series,_10k_chunks,_10k_samples-8     349ms ± 8%     271ms ± 2%  -22.17%  (p=0.000 n=9+10)

name                                                               old alloc/op   new alloc/op   delta
QueryStreamResponseUnmarshal/10_series,_10_chunks,_100_samples-8      143kB ± 0%      47kB ± 0%  -67.38%  (p=0.000 n=10+9)
QueryStreamResponseUnmarshal/10_series,_100_chunks,_100_samples-8     250kB ± 0%      47kB ± 0%  -81.37%  (p=0.000 n=10+10)
QueryStreamResponseUnmarshal/100_series,_1k_chunks,_1k_samples-8     21.5MB ± 0%     5.1MB ± 0%  -76.05%  (p=0.000 n=10+10)
QueryStreamResponseUnmarshal/100_series,_10k_chunks,_1k_samples-8     214MB ± 0%       7MB ± 1%  -96.56%  (p=0.000 n=10+9)
QueryStreamResponseUnmarshal/1k_series,_10k_chunks,_10k_samples-8    2.77GB ± 0%    1.21GB ± 0%  -56.46%  (p=0.000 n=10+8)

name                                                               old allocs/op  new allocs/op  delta
QueryStreamResponseUnmarshal/10_series,_10_chunks,_100_samples-8        221 ± 0%        93 ± 0%  -57.92%  (p=0.000 n=10+10)
QueryStreamResponseUnmarshal/10_series,_100_chunks,_100_samples-8       251 ± 0%        93 ± 0%  -62.95%  (p=0.000 n=10+10)
QueryStreamResponseUnmarshal/100_series,_1k_chunks,_1k_samples-8      3.22k ± 0%     1.24k ± 0%  -61.61%  (p=0.000 n=10+10)
QueryStreamResponseUnmarshal/100_series,_10k_chunks,_1k_samples-8     3.92k ± 0%     1.25k ± 0%  -68.01%  (p=0.000 n=10+10)
QueryStreamResponseUnmarshal/1k_series,_10k_chunks,_10k_samples-8     46.0k ± 0%     26.3k ± 0%  -42.83%  (p=0.000 n=10+10)
```

#### Which issue(s) this PR fixes or relates to

Fixes #3265 

#### Checklist

- [X] Tests updated
- [ ] Documentation added
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`