@ortuman 2022-10-27
FTR. I gave a shot to this new implementation in one of our dev environments (`cortex-dev-01`) and I'd say the difference is at least noticeable.

Below I've included screenshots of the profiles comparing the old and new implementations, both corresponding to `ruler-querier` component:

Before: 
![before-ruler-querier](https://user-images.githubusercontent.com/888899/198269070-fdfea805-dd7d-408d-91a5-110d95bd3315.png)

After:
![after-ruler-querier](https://user-images.githubusercontent.com/888899/198269181-962a07d3-917b-4e65-8684-d5ae5e8f3155.png)

After comparison, we see a reduction of `16.66%` in total number of allocated objects when unmarshaling stream responses.

Worth adding, that memory allocation drop can also be noticed when looking at the GC pause rate.

![Screenshot 2022-10-27 at 13 03 07](https://user-images.githubusercontent.com/888899/198269482-a8e36f91-66c3-4aab-bfe9-bca13c5cc0d2.png)