@pracucci 2022-11-03
Have you also considered adding a correctness test? I would call `QueryStream()` concurrently many times (with a cap to concurrency, e.g. 10 concurrent requests, for a total of 1K requests). Each call to `QueryStream()` should pass a different set of matchers. The mocked ingester should return a different result based on the label matchers (with variable number of series labels and chunks). Then each result of `QueryStream()` should be checked to make sure it matches the expected data.