@pracucci 2023-05-25
My main concern around any optimization like this, is the copying we have to do for the protobuf generated code (`pkg/ingester/client/wrapped.go`) which makes it hard to maintain. I know we're already doing it in other places, and I have concerns too. If benchmarks are good, I will **not** block on this, but longer term we should have a better strategy.