@ortuman 2023-05-29
Following @charleskorn suggestion, I've added a new benchmark named `BenchmarkDistributor_QueryStream` to this PR in order to test the complete execution flow for the `QueryStream` method. After rebasing with the main branch and adapting the code to recent changes, I've observed that the performance of this branch is lower compared to the current implementation.

Results:
```
Distributor_QueryStream/10_metrics,_100_samples-8     162µs ± 3%     219µs ± 1%  +35.41%  (p=0.000 n=10+10)
Distributor_QueryStream/100_series,_1k_samples-8     6.37ms ± 1%   10.64ms ± 0%  +66.90%  (p=0.000 n=10+10)
Distributor_QueryStream/1k_series,_1k_samples-8       198ms ± 1%     301ms ± 6%  +51.85%  (p=0.000 n=9+9)

name                                               old alloc/op   new alloc/op   delta
Distributor_QueryStream/10_metrics,_100_samples-8     304kB ± 0%     398kB ± 0%  +30.73%  (p=0.000 n=10+10)
Distributor_QueryStream/100_series,_1k_samples-8     15.4MB ± 0%    20.3MB ± 0%  +31.32%  (p=0.000 n=10+10)
Distributor_QueryStream/1k_series,_1k_samples-8       320MB ± 0%     443MB ± 0%  +38.48%  (p=0.000 n=10+8)

name                                               old allocs/op  new allocs/op  delta
Distributor_QueryStream/10_metrics,_100_samples-8     4.19k ± 0%     5.42k ± 0%  +29.47%  (p=0.000 n=10+7)
Distributor_QueryStream/100_series,_1k_samples-8       199k ± 0%      260k ± 0%  +30.53%  (p=0.000 n=10+10)
Distributor_QueryStream/1k_series,_1k_samples-8       3.99M ± 0%     5.58M ± 0%  +39.83%  (p=0.000 n=10+8)
```

My assumption here is that the performance gained when unmarshaling the timeseries chunks is subsequently lost due to the need to make a deep copy before returning the result. I understand that the recent changes have outperformed the advantages introduced in this branch, so if you think it's appropriate, you can close it. 

And by the way, great job. 😉